Refactored
Alle itemid'er skulle gerne være string nu (på database forbindelses siden).

Implementering af SQL klassens metoder

+++ src/server/model/item/SaleStrategy/BuyoutStrategy.java
@@ -11,8 +11,7 @@ public class BuyoutStrategy implements SaleStrategy {
 	private double price;
 	private String buyer;
 
-	public BuyoutStrategy(int price)
-	{
+	public BuyoutStrategy(double price) {
 		this.price = price;
 	}
